requirements.txt
Django==3.2.3
psycopg2-binary>=2.8
django-environ==0.4.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-api-key==2.0.0
djangorestframework-camel-case==1.2.0

drf-yasg==1.20.0
drf-yasg-examples==0.1.1

requests==2.25.1
gunicorn==20.0.4
psycopg2-binary==2.8.5